# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Foreman AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crews:
    get:
      operationId: listCrews
      tags: [crews]
      summary: List crews for the current user's company
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of crews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crew"
        "401":
          description: Unauthorized
    post:
      operationId: createCrew
      tags: [crews]
      summary: Create a new crew
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewInput"
      responses:
        "201":
          description: Created crew
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /crews/{crewId}:
    get:
      operationId: getCrew
      tags: [crews]
      summary: Get a crew by ID
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crew
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateCrew
      tags: [crews]
      summary: Update a crew
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewUpdate"
      responses:
        "200":
          description: Updated crew
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteCrew
      tags: [crews]
      summary: Delete a crew
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /crews/{crewId}/members:
    get:
      operationId: listCrewMembers
      tags: [crews]
      summary: List members of a crew
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crew members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewMember"
        "401":
          description: Unauthorized
    post:
      operationId: addCrewMember
      tags: [crews]
      summary: Add a member to a crew
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewMemberInput"
      responses:
        "201":
          description: Crew member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /crews/{crewId}/members/{memberId}:
    patch:
      operationId: updateCrewMember
      tags: [crews]
      summary: Update a crew member
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewMemberUpdate"
      responses:
        "200":
          description: Updated crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: removeCrewMember
      tags: [crews]
      summary: Remove a member from a crew
      parameters:
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  # ──── Projects ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Crew:
      type: object
      required: [id, companyId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        foremanId:
          type: ["integer", "null"]
        foremanName:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    CrewInput:
      type: object
      required: [companyId, name]
      properties:
        companyId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
        foremanId:
          type: integer
    CrewUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        foremanId:
          type: ["integer", "null"]
    CrewMember:
      type: object
      required: [id, crewId, name, trade, createdAt]
      properties:
        id:
          type: integer
        crewId:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        trade:
          type: string
        classification:
          type: ["string", "null"]
        createdAt:
          type: string
    CrewMemberInput:
      type: object
      required: [name, trade]
      properties:
        name:
          type: string
          minLength: 1
        trade:
          type: string
          minLength: 1
        classification:
          type: string
        userId:
          type: integer
    CrewMemberUpdate:
      type: object
      properties:
        name:
          type: string
        trade:
          type: string
        classification:
          type: ["string", "null"]

    # ── Project ────────────────────────────────────────────────────────────
...