# lib/api-zod/src/generated/types/crew.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Foreman AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Crew {
  id: number;
  companyId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  foremanId?: number | null;
  /** @nullable */
  foremanName?: string | null;
  memberCount?: number;
  createdAt: string;
  updatedAt: string;
}